+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Add
+- Selectable regions on Scene window
 - CinemachineRegionsConfiner extension
 - Serializable Region data
 - Cinemachine 2.2.0 dependency